+++ drake/bindings/pybind11/CMakeLists.txt
@@ -7,14 +7,14 @@ else()
 endif()
 
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wno-shadow")
-pybind11_add_module(_rbtree rbtree.cc)
-target_link_libraries(_rbtree PRIVATE drakeRBM drakeMultibodyParsers Eigen3::Eigen)
-install(TARGETS _rbtree DESTINATION ${PYDRAKE_INSTALL_DIR})
+pybind11_add_module(_pydrake_rbtree pydrake_rbtree.cc)
+target_link_libraries(_pydrake_rbtree PRIVATE drakeRBM drakeMultibodyParsers Eigen3::Eigen)
+install(TARGETS _pydrake_rbtree DESTINATION ${PYDRAKE_INSTALL_DIR})
 
-pybind11_add_module(_autodiffutils autodiffutils.cc)
-target_link_libraries(_autodiffutils PRIVATE drakeRBM drakeMultibodyParsers Eigen3::Eigen)
-install(TARGETS _autodiffutils DESTINATION ${PYDRAKE_INSTALL_DIR})
+pybind11_add_module(_pydrake_autodiffutils pydrake_autodiffutils.cc)
+target_link_libraries(_pydrake_autodiffutils PRIVATE drakeRBM drakeMultibodyParsers Eigen3::Eigen)
+install(TARGETS _pydrake_autodiffutils DESTINATION ${PYDRAKE_INSTALL_DIR})
 
-pybind11_add_module(_ik ik.cc)
-target_link_libraries(_ik PRIVATE drakeIK Eigen3::Eigen)
-install(TARGETS _ik DESTINATION ${PYDRAKE_INSTALL_DIR}/solvers)
+pybind11_add_module(_pydrake_ik pydrake_ik.cc)
+target_link_libraries(_pydrake_ik PRIVATE drakeIK Eigen3::Eigen)
+install(TARGETS _pydrake_ik DESTINATION ${PYDRAKE_INSTALL_DIR}/solvers)
